+++ src/views/feature/excel/choiceExport.vue
@@ -158,7 +158,6 @@
         return [
           {
             label: '删除',
-            icon: 'ic:outline-delete-outline',
             onClick: handleDelete.bind(null, record),
             // 根据业务控制是否显示 isShow 和 auth 是并且关系
             ifShow: () => {