+++ src/views/system/user/user.vue
@@ -134,7 +134,6 @@
         actions: [
           {
             label: '删除',
-            icon: 'ic:outline-delete-outline',
             onClick: handleDelete.bind(null, record),
           },
           {